+++ R/tableby.stats.R
@@ -247,6 +247,12 @@ gmeanCI <- function(x, na.rm=TRUE, weights = NULL, conf.level = 0.95, ...) {
   as.tbstat(y, parens = c("(", ")"), sep2 = ", ")
 }
 
+#' @rdname tableby.stats
+#' @export
+Nsigntest <- function(x, na.rm = TRUE, weights = NULL, ...) {
+  if(is.null(weights)) weights <- rep(1, NROW(x))
+  as.countpct(sum(weights*(x != 0), na.rm = na.rm))
+}
 
 ## survival stats
 #' @rdname tableby.stats
@@ -377,9 +383,10 @@ iqr <- function(x, na.rm=TRUE, weights = NULL, ...) {
 ## Count of missings: always show missings
 #' @rdname tableby.stats
 #' @export
-Nmiss <- function(x, na.rm=TRUE, weights = NULL, ...) {
+Nmiss <- function(x, weights = NULL, ...) {
   if(is.null(weights)) weights <- rep(1, NROW(x))
-  as.countpct(sum(weights[is.na(x)]))
+  weights <- weights[is.na(x) | is.na(weights)]
+  as.countpct(sum(weights))
 }
 
 ## Nmiss2 make similar, but in tableby, always keep nmiss,
@@ -393,7 +400,8 @@ Nmiss2 <- Nmiss
 #' @export
 N <- function(x, na.rm=TRUE, weights = NULL, ...) {
   if(is.null(weights)) weights <- rep(1, NROW(x))
-  as.countpct(sum(weights[!is.na(x)]))
+  if(na.rm) weights <- weights[!is.na(x) & !is.na(weights)]
+  as.countpct(sum(weights))
 }
 
 #' @rdname tableby.stats